# Fix: `eslint.verify` should not mutate config argument (fixes #8329)