@ilyavolodin 2017-03-26
@alberto Could you do perf run before and after this change?