@alberto 2017-03-26
@ilyavolodin no noticeable difference:

Before:
```
Loading:
  Load performance Run #1:  174.65021ms
  Load performance Run #2:  168.799739ms
  Load performance Run #3:  178.816711ms
  Load performance Run #4:  177.747853ms
  Load performance Run #5:  182.442309ms

  Load Performance median:  177.747853ms


Single File:
  CPU Speed is 2600 with multiplier 13000000
  Performance Run #1:  4207.118519ms
  Performance Run #2:  4098.461712ms
  Performance Run #3:  3699.8055409999997ms
  Performance Run #4:  3638.281188ms
  Performance Run #5:  3681.332236ms

  Performance budget ok:  3699.8055409999997ms (limit: 5000ms)


Multi Files (0 files):
  CPU Speed is 2600 with multiplier 39000000
  Performance Run #1:  12933.732048ms
  Performance Run #2:  12714.120701ms
  Performance Run #3:  12723.325213ms
  Performance Run #4:  12861.724682ms
  Performance Run #5:  12863.99631ms

  Performance budget ok:  12861.724682ms (limit: 15000ms)
```

After:
```
Loading:
  Load performance Run #1:  174.81977ms
  Load performance Run #2:  179.599493ms
  Load performance Run #3:  183.244189ms
  Load performance Run #4:  181.803644ms
  Load performance Run #5:  165.226383ms

  Load Performance median:  179.599493ms


Single File:
  CPU Speed is 2600 with multiplier 13000000
  Performance Run #1:  3713.787855ms
  Performance Run #2:  3659.128527ms
  Performance Run #3:  3659.855329ms
  Performance Run #4:  3786.704485ms
  Performance Run #5:  4434.773668ms

  Performance budget ok:  3713.787855ms (limit: 5000ms)


Multi Files (0 files):
  CPU Speed is 2600 with multiplier 39000000
  Performance Run #1:  12947.623801ms
  Performance Run #2:  12859.656538ms
  Performance Run #3:  12722.837927ms
  Performance Run #4:  12703.583254ms
  Performance Run #5:  13003.146405ms

  Performance budget ok:  12859.656538ms (limit: 15000ms)
```